refactored to remove controller, this will be a supporting module
